SKALE-3590 move methods to debug interface
