+++ libweb3jsonrpc/skaleStats.json
@@ -2,10 +2,6 @@
 { "name": "skale_stats", "params": [], "order": [], "returns": {}},
 { "name": "skale_nodesRpcInfo", "params": [], "order": [], "returns": {}},
 { "name": "skale_imaInfo", "params": [], "order": [], "returns": {}},
-{ "name": "skale_imaVerifyAndSign", "params": [], "order": [], "returns": {}},
-{ "name": "skale_performanceTrackingStatus", "params": [], "order": [], "returns": {}},
-{ "name": "skale_performanceTrackingStart", "params": [], "order": [], "returns": {}},
-{ "name": "skale_performanceTrackingStop", "params": [], "order": [], "returns": {}},
-{ "name": "skale_performanceTrackingFetch", "params": [], "order": [], "returns": {}}
-];
+{ "name": "skale_imaVerifyAndSign", "params": [], "order": [], "returns": {}}
+]
